requirements.txt
torch>=2.2
transformers>=4.44
datasets>=2.20
tqdm
numpy
matplotlib
pyyaml